requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pdfplumber
python-docx
python-multipart
google-generativeai
numpy
motor
passlib[argon2]
